requirements.txt
yfinance==0.2.36
prophet==1.1.5
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.0
torch==2.0.1
matplotlib==3.7.2
jupyter==1.0.0
python-dotenv==1.0.0
tqdm==4.65.0
streamlit
pandas
numpy
textblob
yfinance
beautifulsoup4
requests
feedparser
plotly
nltk
scikit-learn
xgboost
fredapi

prophet
lime
python-dotenv